refactor(#97): vehicles/utils — bulletproof-react segmentation
git mv:
- transportTypeRef.ts, vehicleFormDefaults.ts → utils/
- projection/vehicleSortValue{,.test}.ts, vehicleUrlParams.ts → utils/

External call sites updated: VehicleListCell, VehicleCreatePage.
