+++ src/data/vehicle-types/cells/VehicleListCell.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Box, Chip, Link } from '@mui/material';
 import { Link as RouterLink } from 'react-router-dom';
-import { vehicleSelectedHref } from '../../vehicles/projection/vehicleUrlParams.ts';
+import { vehicleSelectedHref } from '../../vehicles/utils/vehicleUrlParams.ts';
 import type { Vehicle } from '../vehicleTypeTypes.ts';
 
 const MAX_VISIBLE = 5;
