+++ src/data/vehicles/projection/useVehicles.ts
@@ -4,7 +4,7 @@ import { useAuth } from '../../../auth/authUtils.ts';
 import type { Order } from '../../../components/data/dataTableTypes.ts';
 import type { VehicleGQLShaped, VehicleColumnKey } from '../types/vehicleGqlShaped.ts';
 import { fetchVehiclesAndApply } from './fetchVehiclesAndApply.ts';
-import { compareVehicles } from './vehicleSortValue.ts';
+import { compareVehicles } from '../utils/vehicleSortValue.ts';
 
 /**
  * Data hook for the `/vehicles` list. Mirrors `useVehicleTypes` — single fetch
