+++ src/data/vehicles/projection/vehicleViewConfig.tsx
@@ -5,9 +5,9 @@ import { useDataViewSearch } from '../../../hooks/useDataViewSearch.ts';
 import { useDataViewTableLogic } from '../../../hooks/useDataViewTableLogic.ts';
 import DataPageContent from '../../../components/data/DataPageContent.tsx';
 import TransportModeIcon from '../../../components/icons/TransportModeIcon.tsx';
-import { getVehicleSortValue } from './vehicleSortValue.ts';
+import { getVehicleSortValue } from '../utils/vehicleSortValue.ts';
 import { useVehicleUrlSelection } from './useVehicleUrlSelection.tsx';
-import { vehicleSelectedHref } from './vehicleUrlParams.ts';
+import { vehicleSelectedHref } from '../utils/vehicleUrlParams.ts';
 import { transportModeFilters } from '../../netex/transportMode.ts';
 import type { ColumnDefinition } from '../../../components/data/dataTableTypes.ts';
 import {
