+++ src/data/vehicles/utils/vehicleSortValue.test.ts
