+++ src/data/vehicles/utils/vehicleSortValue.ts
