+++ src/data/vehicles/vehicleFormState.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { edit, hydrate, initialFormState, isDirty, type FormState } from './vehicleFormState';
-import { BLANK_FORM, MISSING_MODEL } from './vehicleFormDefaults';
+import { BLANK_FORM, MISSING_MODEL } from './utils/vehicleFormDefaults';
 import type { Vehicle } from './types/Vehicle';
 import type { VehicleEditFormValue } from './VehicleEditForm';
